Refactor the new SelectInterpreter task.
Basically just breaks up the big execute() method
into several helpers.

This is for two reasons:

- Makes it easier to test invalidation (see below).
- Will enable selecting interpreters separately for
  each target and its deps (for use by the PythonEval task)
  as well as the current functionality of selecting a single
  interpreter for the entire closure. This will be implemented
  in a future change.

Note that testing invalidation isn't an idle idea - I noticed
a bug in the custom fingerprint strategy that was mixing in
the target payload for no good reason.  This change fixes
that bug and adds a test for it.